requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.26.0
flask-restx==1.3.0
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
marshmallow==3.14.0
bcrypt==3.2.0
psycopg2==2.9.1
python-dotenv==0.19.1
pylint==2.11.1
MarkupSafe<3.0.0,>=2.1.1
urllib3<1.27,>=1.21.1
certifi>=2017.4.17
charset-normalizer~=2.0.0
idna<4,>=2.5
aniso8601>=0.82
jsonschema==4.22.0
sentry_sdk
flask_jwt_extended
cryptography
xlrd
pandas